BUG: Procrustes-aligned output has parametric scalar data if enabled.
The code to copy the array values into scalars on the VTK output should probably be deduplicated - it is already duplicated twice for the `_SPHARM.vtk` and `_SPHARM_ellalign.vtk` outputs. In the interested of time I've just duplicated it again for this `_SPHARM_procalign.vtk` output...

For the SPHARM-PDM modularization refactor this should come for free since these steps are implemented as VTK filters and the scalars are automatically passed through... so I don't think this tech debt is too much.